pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "plex-poster-healer"
version = "0.1.0"
description = "Scan and heal broken Plex artwork"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "boto3>=1.34.0",
  "jinja2>=3.1.0",
  "Pillow>=10.0.0",
  "plexapi>=4.15.0",
  "pydantic>=2.6.0",
  "python-dotenv>=1.0.0",
  "PyYAML>=6.0.1",
  "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
plex-poster-healer = "plex_poster_healer.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
